@@ -6,10 +6,10 @@ build-backend = "hatchling.build"
 name = "schemathesis"
 version = "3.21.2"
 description = "Property-based testing framework for Open API and GraphQL based apps"
-keywords = ["pytest", "hypothesis", "openapi", "swagger", "graphql", "testing"]
+keywords = ["pytest", "hypothesis", "openapi", "graphql", "testing"]
 authors = [{ name = "Dmitry Dygalo", email = "dadygalo@gmail.com" }]
 maintainers = [{ name = "Dmitry Dygalo", email = "dadygalo@gmail.com" }]
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 license = "MIT"
 readme = "README.md"
 include = ["src/schemathesis/py.typed"]
@@ -22,7 +22,6 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
@@ -39,7 +38,6 @@ dependencies = [
     "hypothesis>=6.31.6,<7",
     "hypothesis_graphql>=0.9.0,<1",
     "hypothesis_jsonschema>=0.22.1,<0.23",
-    "importlib_metadata>=1.1,!=3.8,<5; python_version<'3.8'",
     "jsonschema>=4.3.2,<5.0",
     "junit-xml>=1.9,<2.0",
     "pyrate-limiter>=2.10,<3.0",
@@ -51,7 +49,6 @@ dependencies = [
     "starlette-testclient==0.2.0",
     "tomli-w>=1.0.0,<2.0",
     "tomli>=2.0.1,<3.0",
-    "typing-extensions>=3.7,<5",
     "werkzeug>=0.16.0,<4",
     "yarl>=1.5,<2.0"
 ]
@@ -149,7 +146,7 @@ ignore = [
     "D213",  # Multiline summary second line
     "D401",  # Imperative mood
 ]
-target-version = "py37"
+target-version = "py38"
 
 # Ignore `F401` (unused import) in all `__init__.py` & `_compat.py` files.
 [tool.ruff.per-file-ignores]
@@ -161,4 +158,4 @@ skip-magic-trailing-comma = false
 
 [tool.ruff.isort]
 known-first-party = ["schemathesis"]
-known-third-party = ["_pytest", "aiohttp", "click", "curlify", "fastapi", "flask", "graphene", "graphql", "graphql_server", "hypothesis", "hypothesis_graphql", "hypothesis_jsonschema", "jsonschema", "junit_xml", "packaging", "pydantic", "pytest", "pytest_subtests", "requests", "starlette", "typing_extensions", "urllib3", "werkzeug", "yaml", "yarl"]
+known-third-party = ["_pytest", "aiohttp", "click", "curlify", "fastapi", "flask", "graphene", "graphql", "graphql_server", "hypothesis", "hypothesis_graphql", "hypothesis_jsonschema", "jsonschema", "junit_xml", "packaging", "pydantic", "pytest", "pytest_subtests", "requests", "starlette", "urllib3", "werkzeug", "yaml", "yarl"]
