@@ -3,7 +3,7 @@ import os
 import shutil
 from dataclasses import dataclass, field
 from queue import Queue
-from typing import List, Optional, Union, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 from ..code_samples import CodeSampleStyle
 from ..runner.serialization import SerializedTestResult
@@ -21,7 +21,7 @@ class ServiceReportContext:
 @dataclass
 class FileReportContext:
     queue: Queue
-    filename: Optional[str] = None
+    filename: str | None = None
 
 
 @dataclass
@@ -29,21 +29,21 @@ class ExecutionContext:
     """Storage for the current context of the execution."""
 
     hypothesis_settings: hypothesis.settings
-    hypothesis_output: List[str] = field(default_factory=list)
+    hypothesis_output: list[str] = field(default_factory=list)
     workers_num: int = 1
-    rate_limit: Optional[str] = None
+    rate_limit: str | None = None
     show_errors_tracebacks: bool = False
-    wait_for_schema: Optional[float] = None
+    wait_for_schema: float | None = None
     validate_schema: bool = True
     operations_processed: int = 0
     # It is set in runtime, from a `Initialized` event
-    operations_count: Optional[int] = None
+    operations_count: int | None = None
     current_line_length: int = 0
     terminal_size: os.terminal_size = field(default_factory=shutil.get_terminal_size)
-    results: List[SerializedTestResult] = field(default_factory=list)
-    cassette_path: Optional[str] = None
-    junit_xml_file: Optional[str] = None
+    results: list[SerializedTestResult] = field(default_factory=list)
+    cassette_path: str | None = None
+    junit_xml_file: str | None = None
     is_interrupted: bool = False
     verbosity: int = 0
     code_sample_style: CodeSampleStyle = CodeSampleStyle.default()
-    report: Optional[Union[ServiceReportContext, FileReportContext]] = None
+    report: ServiceReportContext | FileReportContext | None = None
