@@ -30,46 +30,46 @@ if TYPE_CHECKING:
 
 @deprecated_function(removed_in="4.0", replacement="schemathesis.runner.from_schema")
 def prepare(
-    schema_uri: Union[str, Dict[str, Any]],
+    schema_uri: str | dict[str, Any],
     *,
     # Runtime behavior
-    checks: Optional[Iterable[CheckFunction]] = None,
-    data_generation_methods: Tuple[DataGenerationMethod, ...] = DEFAULT_DATA_GENERATION_METHODS,
-    max_response_time: Optional[int] = None,
+    checks: Iterable[CheckFunction] | None = None,
+    data_generation_methods: tuple[DataGenerationMethod, ...] = DEFAULT_DATA_GENERATION_METHODS,
+    max_response_time: int | None = None,
     targets: Iterable[Target] = DEFAULT_TARGETS,
     workers_num: int = 1,
-    seed: Optional[int] = None,
+    seed: int | None = None,
     exit_first: bool = False,
     dry_run: bool = False,
     store_interactions: bool = False,
-    stateful: Optional[Stateful] = None,
+    stateful: Stateful | None = None,
     stateful_recursion_limit: int = DEFAULT_STATEFUL_RECURSION_LIMIT,
     # Schema loading
     loader: Callable = oas_loaders.from_uri,
-    base_url: Optional[str] = None,
-    auth: Optional[Tuple[str, str]] = None,
-    auth_type: Optional[str] = None,
-    headers: Optional[Dict[str, str]] = None,
-    request_timeout: Optional[int] = None,
-    request_tls_verify: Union[bool, str] = True,
-    request_cert: Optional[RequestCert] = None,
-    endpoint: Optional[Filter] = None,
-    method: Optional[Filter] = None,
-    tag: Optional[Filter] = None,
-    operation_id: Optional[Filter] = None,
-    app: Optional[str] = None,
+    base_url: str | None = None,
+    auth: tuple[str, str] | None = None,
+    auth_type: str | None = None,
+    headers: dict[str, str] | None = None,
+    request_timeout: int | None = None,
+    request_tls_verify: bool | str = True,
+    request_cert: RequestCert | None = None,
+    endpoint: Filter | None = None,
+    method: Filter | None = None,
+    tag: Filter | None = None,
+    operation_id: Filter | None = None,
+    app: str | None = None,
     validate_schema: bool = True,
     skip_deprecated_operations: bool = False,
-    force_schema_version: Optional[str] = None,
+    force_schema_version: str | None = None,
     count_operations: bool = True,
     # Hypothesis-specific configuration
-    hypothesis_deadline: Optional[Union[int, NotSet]] = None,
-    hypothesis_derandomize: Optional[bool] = None,
-    hypothesis_max_examples: Optional[int] = None,
-    hypothesis_phases: Optional[List[hypothesis.Phase]] = None,
-    hypothesis_report_multiple_bugs: Optional[bool] = None,
-    hypothesis_suppress_health_check: Optional[List[hypothesis.HealthCheck]] = None,
-    hypothesis_verbosity: Optional[hypothesis.Verbosity] = None,
+    hypothesis_deadline: int | NotSet | None = None,
+    hypothesis_derandomize: bool | None = None,
+    hypothesis_max_examples: int | None = None,
+    hypothesis_phases: list[hypothesis.Phase] | None = None,
+    hypothesis_report_multiple_bugs: bool | None = None,
+    hypothesis_suppress_health_check: list[hypothesis.HealthCheck] | None = None,
+    hypothesis_verbosity: hypothesis.Verbosity | None = None,
 ) -> Generator[events.ExecutionEvent, None, None]:
     """Prepare a generator that will run test cases against the given API definition."""
     from ..checks import DEFAULT_CHECKS
@@ -124,7 +124,7 @@ def prepare(
     )
 
 
-def validate_loader(loader: Callable, schema_uri: Union[str, Dict[str, Any]]) -> None:
+def validate_loader(loader: Callable, schema_uri: str | dict[str, Any]) -> None:
     """Sanity checking for input schema & loader."""
     if loader not in (
         oas_loaders.from_uri,
@@ -149,34 +149,34 @@ def validate_loader(loader: Callable, schema_uri: Union[str, Dict[str, Any]]) ->
 
 def execute_from_schema(
     *,
-    schema_uri: Union[str, Dict[str, Any]],
+    schema_uri: str | dict[str, Any],
     loader: Callable = oas_loaders.from_uri,
-    base_url: Optional[str] = None,
-    endpoint: Optional[Filter] = None,
-    method: Optional[Filter] = None,
-    tag: Optional[Filter] = None,
-    operation_id: Optional[Filter] = None,
-    app: Optional[str] = None,
+    base_url: str | None = None,
+    endpoint: Filter | None = None,
+    method: Filter | None = None,
+    tag: Filter | None = None,
+    operation_id: Filter | None = None,
+    app: str | None = None,
     validate_schema: bool = True,
     skip_deprecated_operations: bool = False,
-    force_schema_version: Optional[str] = None,
+    force_schema_version: str | None = None,
     checks: Iterable[CheckFunction],
-    data_generation_methods: Tuple[DataGenerationMethod, ...] = DEFAULT_DATA_GENERATION_METHODS,
-    max_response_time: Optional[int] = None,
+    data_generation_methods: tuple[DataGenerationMethod, ...] = DEFAULT_DATA_GENERATION_METHODS,
+    max_response_time: int | None = None,
     targets: Iterable[Target],
     workers_num: int = 1,
     hypothesis_settings: hypothesis.settings,
-    auth: Optional[RawAuth] = None,
-    auth_type: Optional[str] = None,
-    headers: Optional[Dict[str, Any]] = None,
-    request_timeout: Optional[int] = None,
-    request_tls_verify: Union[bool, str] = True,
-    request_cert: Optional[RequestCert] = None,
-    seed: Optional[int] = None,
+    auth: RawAuth | None = None,
+    auth_type: str | None = None,
+    headers: dict[str, Any] | None = None,
+    request_timeout: int | None = None,
+    request_tls_verify: bool | str = True,
+    request_cert: RequestCert | None = None,
+    seed: int | None = None,
     exit_first: bool = False,
     dry_run: bool = False,
     store_interactions: bool = False,
-    stateful: Optional[Stateful] = None,
+    stateful: Stateful | None = None,
     stateful_recursion_limit: int = DEFAULT_STATEFUL_RECURSION_LIMIT,
     count_operations: bool = True,
 ) -> Generator[events.ExecutionEvent, None, None]:
@@ -234,26 +234,26 @@ def execute_from_schema(
 
 
 def load_schema(
-    schema_uri: Union[str, Dict[str, Any]],
+    schema_uri: str | dict[str, Any],
     *,
-    base_url: Optional[str] = None,
+    base_url: str | None = None,
     loader: Callable = oas_loaders.from_uri,
     app: Any = None,
     validate_schema: bool = True,
     skip_deprecated_operations: bool = False,
-    data_generation_methods: Tuple[DataGenerationMethod, ...] = DEFAULT_DATA_GENERATION_METHODS,
-    force_schema_version: Optional[str] = None,
-    request_tls_verify: Union[bool, str] = True,
-    request_cert: Optional[RequestCert] = None,
+    data_generation_methods: tuple[DataGenerationMethod, ...] = DEFAULT_DATA_GENERATION_METHODS,
+    force_schema_version: str | None = None,
+    request_tls_verify: bool | str = True,
+    request_cert: RequestCert | None = None,
     # Network request parameters
-    auth: Optional[Tuple[str, str]] = None,
-    auth_type: Optional[str] = None,
-    headers: Optional[Dict[str, str]] = None,
+    auth: tuple[str, str] | None = None,
+    auth_type: str | None = None,
+    headers: dict[str, str] | None = None,
     # Schema filters
-    endpoint: Optional[Filter] = None,
-    method: Optional[Filter] = None,
-    tag: Optional[Filter] = None,
-    operation_id: Optional[Filter] = None,
+    endpoint: Filter | None = None,
+    method: Filter | None = None,
+    tag: Filter | None = None,
+    operation_id: Filter | None = None,
 ) -> BaseSchema:
     """Load schema via specified loader and parameters."""
     loader_options = {
@@ -306,24 +306,24 @@ def load_schema(
 def from_schema(
     schema: BaseSchema,
     *,
-    checks: Optional[Iterable[CheckFunction]] = None,
-    max_response_time: Optional[int] = None,
+    checks: Iterable[CheckFunction] | None = None,
+    max_response_time: int | None = None,
     targets: Iterable[Target] = DEFAULT_TARGETS,
     workers_num: int = 1,
-    hypothesis_settings: Optional[hypothesis.settings] = None,
-    auth: Optional[RawAuth] = None,
-    auth_type: Optional[str] = None,
-    headers: Optional[Dict[str, Any]] = None,
-    request_timeout: Optional[int] = None,
-    request_tls_verify: Union[bool, str] = True,
-    request_cert: Optional[RequestCert] = None,
-    seed: Optional[int] = None,
+    hypothesis_settings: hypothesis.settings | None = None,
+    auth: RawAuth | None = None,
+    auth_type: str | None = None,
+    headers: dict[str, Any] | None = None,
+    request_timeout: int | None = None,
+    request_tls_verify: bool | str = True,
+    request_cert: RequestCert | None = None,
+    seed: int | None = None,
     exit_first: bool = False,
-    max_failures: Optional[int] = None,
-    started_at: Optional[str] = None,
+    max_failures: int | None = None,
+    started_at: str | None = None,
     dry_run: bool = False,
     store_interactions: bool = False,
-    stateful: Optional[Stateful] = None,
+    stateful: Stateful | None = None,
     stateful_recursion_limit: int = DEFAULT_STATEFUL_RECURSION_LIMIT,
     count_operations: bool = True,
 ) -> BaseRunner:
@@ -473,14 +473,14 @@ def from_schema(
 
 
 def prepare_hypothesis_settings(
-    database: Optional[str] = None,
-    deadline: Optional[Union[int, NotSet]] = None,
-    derandomize: Optional[bool] = None,
-    max_examples: Optional[int] = None,
-    phases: Optional[List[hypothesis.Phase]] = None,
-    report_multiple_bugs: Optional[bool] = None,
-    suppress_health_check: Optional[List[hypothesis.HealthCheck]] = None,
-    verbosity: Optional[hypothesis.Verbosity] = None,
+    database: str | None = None,
+    deadline: int | NotSet | None = None,
+    derandomize: bool | None = None,
+    max_examples: int | None = None,
+    phases: list[hypothesis.Phase] | None = None,
+    report_multiple_bugs: bool | None = None,
+    suppress_health_check: list[hypothesis.HealthCheck] | None = None,
+    verbosity: hypothesis.Verbosity | None = None,
 ) -> hypothesis.settings:
     import hypothesis
     from hypothesis.database import DirectoryBasedExampleDatabase, InMemoryExampleDatabase
