@@ -1,5 +1,5 @@
 from __future__ import annotations
-from typing import Dict, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 
 from ...exceptions import UsageError
@@ -8,7 +8,7 @@ if TYPE_CHECKING:
     import graphql
     from hypothesis import strategies as st
 
-CUSTOM_SCALARS: Dict[str, st.SearchStrategy[graphql.ValueNode]] = {}
+CUSTOM_SCALARS: dict[str, st.SearchStrategy[graphql.ValueNode]] = {}
 
 
 def scalar(name: str, strategy: st.SearchStrategy[graphql.ValueNode]) -> None:
