@@ -1,5 +1,5 @@
 from __future__ import annotations
-from typing import Optional, Union, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 from ..types import RawAuth
 
@@ -7,7 +7,7 @@ if TYPE_CHECKING:
     from requests.auth import HTTPDigestAuth
 
 
-def get_requests_auth(auth: Optional[RawAuth], auth_type: Optional[str]) -> Optional[Union[HTTPDigestAuth, RawAuth]]:
+def get_requests_auth(auth: RawAuth | None, auth_type: str | None) -> HTTPDigestAuth | RawAuth | None:
     from requests.auth import HTTPDigestAuth
 
     if auth and auth_type == "digest":
