@@ -1,4 +1,4 @@
-from ._compat import metadata
+from importlib import metadata
 from .types import NotSet
 
 try:
